Configure SASL for remote mail submission
SASL allows remote hosts to authenticate and send email without being
part of `mynetworks`.

In this case, I opted to use the "sasldb" authentication backend. This unfortunately requires manual setup (for
now).

With this config, SASL submission and auth is only allowed over TLS to
ensure that plaintext creds are encrypted in transit.

In addition, I have:

* Fixed the Postfix main.cf config (default_transport!)
* Used @localhost for local aliases
* Added CA cert path for Postfix client and server

Closes #22.