Server: Handle case of filter/client retries
If a retry comes in and we've already processed a subset of the
attachments, then there already will be an entry for the email in the
cache. To handle this, we check the cache first for any incoming email,
and if an entry is found, we restore the attachment count to allow
processing to work as expected.